Revert "OPENEUROPA-1328: Improve code and behat tests."
This reverts commit bab8719bcc2aad0d417f7c144e1219526d7a33ff.
