+++ .drone.yml
@@ -27,56 +27,50 @@ pipeline:
     volumes:
       - /cache:/cache
     commands:
+      # @todo remove "composer install" step once the following issue is fixed.
+      # @link https://webgate.ec.europa.eu/CITnet/jira/browse/OPENEUROPA-1234
       - composer install --ansi --no-suggest --no-progress
-
-  composer-update-prefer-lowest:
+      - composer update --prefer-lowest --prefer-stable --ansi --no-suggest --no-progress
+    when:
+      matrix:
+        COMPOSER_BOUNDARY: lowest
+  composer-install-highest:
     group: prepare
     image: fpfis/httpd-php-dev:7.1
     volumes:
       - /cache:/cache
     commands:
-      # @todo remove "composer install" step once the following issue is fixed.
-      # @link https://webgate.ec.europa.eu/CITnet/jira/browse/OPENEUROPA-1234
-      - composer update --prefer-lowest --prefer-stable --ansi --no-suggest --no-progress
+      - composer install --ansi --no-suggest --no-progress
     when:
       matrix:
-        COMPOSER_BOUNDARY: lowest
+        COMPOSER_BOUNDARY: highest
 
   site-install:
     image: fpfis/httpd-php-dev:7.1
     commands:
       - ./vendor/bin/run drupal:site-install
 
-  test-grumphp:
+  grumphp:
     group: test
     image: fpfis/httpd-php-dev:7.1
     commands:
       - ./vendor/bin/grumphp run
 
-  test-phpunit:
+  phpunit:
     group: test
     image: fpfis/httpd-php-dev:7.1
     commands:
       - ./vendor/bin/phpunit
 
-  test-behat:
+  behat:
     group: test
     image: fpfis/httpd-php-dev:7.1
     commands:
-      # Reset permission since installation runs as root. @todo Fix this on drone.
-      - ./vendor/bin/drush -y config-set system.performance css.preprocess 0
-      - ./vendor/bin/drush -y config-set system.performance js.preprocess 0
-      - chmod -R 777 build
-      # Run behat tests.
+      # Workaround for fixing problem with files permission inside docker
+      # @todo remove first 2 lines after docker image release which contain fix of this problem
+      - chmod -R 777 build/sites/default/files
+      - ./vendor/bin/drush cr
       - ./vendor/bin/behat
-
-  debug:
-    image: fpfis/httpd-php-dev:7.1
-    commands:
-      - ./vendor/bin/drush ws --count 500
-    when:
-      status: failure
-
 matrix:
   COMPOSER_BOUNDARY:
     - lowest
