+++ composer.json
@@ -1,65 +1,58 @@
 {
-  "name": "openeuropa/oe_media",
-  "description": "Media features for the OpenEuropa project.",
-  "type": "drupal-module",
-  "license": "EUPL-1.2",
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "require": {
-    "drupal/entity_browser": "^2.0",
-    "drupal/inline_entity_form": "^1.0",
-    "drupal/media_avportal": "1.x-dev",
-    "openeuropa/drupal-core-require": "^8.6",
-    "php": "^7.1"
-  },
-  "require-dev": {
-    "composer/installers": "~1.5",
-    "drush/drush": "~9.0",
-    "drupal-composer/drupal-scaffold": "^2.5.2",
-    "drupal/ctools": "^3.0",
-    "drupal/drupal-extension": "^4.0.0@alpha",
-    "nikic/php-parser": "~3.0",
-    "openeuropa/code-review": "~1.0.0-alpha4",
-    "openeuropa/drupal-core-require-dev": "~8.6",
-    "openeuropa/task-runner": "~1.0.0-beta2",
-    "phpunit/phpunit": "~5.5||~6.0",
-    "symfony/browser-kit": "~3.0||~4.0"
-  },
-  "scripts": {
-    "drupal-scaffold": "DrupalComposer\\DrupalScaffold\\Plugin::scaffold",
-    "post-install-cmd": "./vendor/bin/run drupal:site-setup",
-    "post-update-cmd": "./vendor/bin/run drupal:site-setup"
-  },
-  "repositories": {
-    "drupal": {
-      "type": "composer",
-      "url": "https://packages.drupal.org/8"
+    "name": "openeuropa/oe_media",
+    "description": "Media features for the OpenEuropa project.",
+    "type": "drupal-module",
+    "license": "EUPL-1.2",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "require": {
+        "php": "^7.1",
+        "drupal/core": "^8.6",
+        "drupal/entity_browser": "^2.0",
+        "drupal/inline_entity_form": "^1.0",
+        "drupal/media_avportal": "1.x-dev"
+    },
+    "require-dev": {
+        "composer/installers": "~1.5",
+        "cweagans/composer-patches": "^1.6.5",
+        "drush/drush": "~9.0@stable",
+        "drupal-composer/drupal-scaffold": "^2.5.2",
+        "drupal/ctools": "^3.0",
+        "openeuropa/code-review": "^1.0.0-alpha4",
+        "openeuropa/task-runner": "~1.0.0-beta2",
+        "openeuropa/drupal-core-require-dev": "^8.6",
+        "drupal/drupal-extension": "^4.0.0@alpha",
+        "phpunit/phpunit": "~5.5||~6.0",
+        "symfony/browser-kit": "~3.0||~4.0"
+    },
+    "scripts": {
+        "drupal-scaffold": "DrupalComposer\\DrupalScaffold\\Plugin::scaffold",
+        "post-install-cmd": "./vendor/bin/run drupal:site-setup",
+        "post-update-cmd": "./vendor/bin/run drupal:site-setup"
+    },
+    "repositories": {
+        "drupal": {
+            "type": "composer",
+            "url": "https://packages.drupal.org/8"
+        }
+    },
+    "autoload": {
+      "psr-4": {
+        "Drupal\\oe_media\\": "./src/"
+      }
+    },
+    "autoload-dev": {
+      "psr-4": {
+        "Drupal\\Tests\\oe_media\\": "./tests/"
+      }
+    },
+    "extra": {
+      "composer-exit-on-patch-failure": true,
+      "enable-patching": true,
+      "installer-paths": {
+        "build/core": ["type:drupal-core"],
+        "build/profiles/contrib/{$name}": ["type:drupal-profile"],
+        "build/modules/contrib/{$name}": ["type:drupal-module"]
+      }
     }
-  },
-  "autoload": {
-    "psr-4": {
-      "Drupal\\oe_media\\": "./src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Drupal\\Tests\\oe_media\\": "./tests/"
-    }
-  },
-  "extra": {
-    "composer-exit-on-patch-failure": true,
-    "enable-patching": true,
-    "installer-paths": {
-      "build/core": ["type:drupal-core"],
-      "build/profiles/contrib/{$name}": ["type:drupal-profile"],
-      "build/modules/contrib/{$name}": ["type:drupal-module"],
-      "build/themes/contrib/{$name}": ["type:drupal-theme"]
-    }
-  },
-  "config": {
-    "sort-packages": true,
-    "platform": {
-      "php": "7.1.9"
-    }
-  }
 }
