+++ docker-compose.yml.dist
@@ -18,8 +18,9 @@ services:
     image: percona/percona-server:5.6
     environment:
       MYSQL_ALLOW_EMPTY_PASSWORD: "yes"
-    # ports:
-    #  - 3306:3306
+#    ports:
+#      - 3306:3306
+
   # If you would like to see what is going on you can run the following on your host:
   # docker run --rm -p 4444:4444 -p 5900:5900 --network="host" selenium/standalone-chrome-debug:latest
   # Newer version of this image might run into this issue:
