+++ grumphp.yml.dist
@@ -11,12 +11,3 @@ parameters:
     - inc
     - theme
     - install
-
-  extra_tasks:
-    phpparser:
-      ignore_patterns: %tasks.phpcs.ignore_patterns%
-      visitors:
-        declare_strict_types: ~
-      triggered_by: %tasks.phpcs.triggered_by%
-  extensions:
-    - OpenEuropa\CodeReview\ExtraTasksExtension
