+++ modules/oe_media_demo/oe_media_demo.info.yml
@@ -7,3 +7,4 @@ dependencies:
   - drupal:media
   - oe_media:oe_media
   - oe_media_avportal:oe_media_avportal
+
