+++ tests/features/av_portal.feature
@@ -4,11 +4,9 @@ Feature: AV Portal.
   As a site editor
   I want to create and reference AV Portal media entities.
 
-  Background:
-    Given I am logged in as a user with the "create oe_media_demo content,create av_portal_video media" permissions
-
-  @demo
+  @av_portal
   Scenario: Create and reference an AV Portal video
+    Given I am logged in as a user with the "create oe_media_demo content,create av_portal_video media" permission
     And I visit "media/add/av_portal_video"
     And I fill in "Media AV Portal Video" with "https://ec.europa.eu/avservices/video/player.cfm?sitelang=en&ref=I-162747"
     And I press "Save"
@@ -17,6 +15,5 @@ Feature: AV Portal.
     And I reference the AV Portal media "Midday press briefing from 25/10/2018"
     And I press "Save"
     Then I should see the AV Portal video "Midday press briefing from 25/10/2018"
-
-    # @todo: Cleanup of the media entity using a tag @afterscenario
+    # Cleanup of the media entity.
     And I remove the media "Midday press briefing from 25/10/2018"
